Add support for PREEMPT_RT kernels
Add a series of #ifdefs in gckOS_WaitSignal() to support RT kernels,
where the completion mechanism has been changed to use swait instead of
complex wait queues.

This is based on a patch previously provided to meta-fsl-arm by
Jacob Kroon:
https://github.com/Freescale/meta-freescale/commit/5ebb4e1950a49c7e1f04ee86ecb0d1929b4ebfc7

In order to be able to load the RT version of the module, the kernel
itself must be patched to export the swait_prepare_locked() and
swait_finish_locked() functions.

Signed-off-by: Dominic Sacré <dominic.sacre@gmx.de>