Make _GetProcessID fetch PIDs from the top-level PID namespace.
This ensures that the returned process ID will uniquely identify a
process, regardless of which PID namespace a process belongs to.

Previously, it was possible (even likely) for _GetProcessID to return
the same value for multiple processes running in distinct PID
namespaces.  This would cause eglGetDisplay to return EGL_NO_DISPLAY
for the second and subsequent processes that the driver incorrectly
determines to have the same process ID.